split agent-schema.ts into heartbeat, swarm, and workspace-git schemas
